Add authentication to routes

+++ routes/advertisements.js
@@ -1,12 +1,16 @@
+const auth = require("../middleware/auth");
+const hasPrivilege = require("../middleware/hasPrivilege");
+
 const { Advertisement } = require('../models/advertisement');
 const { User } = require('../models/user')
+
 const express = require('express');
 
 
 const mongoose = require('mongoose');
 const router = express.Router();
 
-router.get('/', async (req,res)=>{
+router.get('/',auth, async (req,res)=>{
     const advertisement = await Advertisement.find().populate('owner');
     if(!advertisement){
         res.status(500).json({
@@ -14,10 +18,10 @@ router.get('/', async (req,res)=>{
             message: "empty advertisement"
         })
     }
-    res.send(advertisement);
+    res.send({advertisement});
 })
 
-router.get('/:id', async (req,res)=>{
+router.get('/:id',auth, async (req,res)=>{
     if(!mongoose.isValidObjectId(req.params.id)){
         res.status(400).send({ message:'Invaild ID'})
     }
@@ -29,17 +33,17 @@ router.get('/:id', async (req,res)=>{
             message: 'the advertisement with given ID was not found'
         })
     }
-    res.status(200).send(advertisement);
+    res.status(200).send({advertisement});
 })
 
-router.post('/',async (req,res)=>{
+router.post('/', auth, async (req,res)=>{
 
     const advertisement =  new Advertisement({
         images: req.body.images,
         address: req.body.address,
         price: req.body.price,
         internet: req.body.internet,
-        owner: req.body.owner,
+        owner: req.body.owner,          //! Get Id of the logged in user .. make it required
         publishedAt: Date.now(),
         apartmentArea: req.body.apartmentArea,
         noOfRooms: req.body.noOfRooms,
@@ -54,10 +58,13 @@ router.post('/',async (req,res)=>{
     return res.status(500).send({ message: ' Error in Creating Advertisement.' });
 })
 
-router.patch('/:id',  async (req,res)=>{
+router.patch('/:id', auth,  async (req,res)=>{
     if(!mongoose.isValidObjectId(req.params.id)){
         res.status(400).send({ message:'Invaild ID'})
     }
+
+    const has_privilege = await hasPrivilege(req, req.params.id);
+    if(!has_privilege) return res.send({message: "You don't have the privilege to perform this action."})
     
     const advertisement = await Advertisement.findByIdAndUpdate(
         req.params.id,
@@ -66,7 +73,7 @@ router.patch('/:id',  async (req,res)=>{
             address: req.body.address,
             price: req.body.price,
             internet: req.body.internet,
-            owner: req.body.owner,
+            owner: req.body.owner,           //! Get Id of the logged in user .. make it required
             publishedAt: Date.now(),
             apartmentArea: req.body.apartmentArea,
             noOfRooms: req.body.noOfRooms,
@@ -77,13 +84,17 @@ router.patch('/:id',  async (req,res)=>{
     if (!advertisement) {
         return res.status(404).send({message:"the advertisement cannot be updated!"});
     }
-    res.send(advertisement);
+    res.send({advertisement});
 })
 
-router.delete('/:id', async (req,res)=>{
+router.delete('/:id', auth, async (req,res)=>{
     if(!mongoose.isValidObjectId(req.params.id)){
         res.status(400).send({ message:'Invaild ID'})
     }
+
+    const has_privilege = await hasPrivilege(req, req.params.id);
+    if(!has_privilege) return res.send({message: "You don't have the privilege to perform this action."})
+
     const advertisement = await Advertisement.findByIdAndRemove(req.params.id)
     if(!advertisement){
         res.status(500).json({
